Create README.md
# Diamond Generator Prank

This is a fun prank app that simulates generating in-game currencies like Free Fire diamonds, Robux, and V-Bucks. The user will enter their ID, confirm it, and then see a message saying, "Success! Your diamonds will be sent in 10 years! ⏳".

## How to Use
1. Click on one of the buttons to generate Diamonds, Robux, or V-Bucks.
2. Enter your Player ID.
3. Confirm your ID.
4. Enjoy the prank message at the end!

This is just for fun, no actual in-game currency is generated!